server/requirements.txt
#web Framework
fastapi
uvicorn[standard]
python-multipart


#lanchain ecosystem
langchain
langchain-community
langchain-core
langchain-groq


#vectordb
pinecone

#embeddings
langchain-google-genai

#pdf parsing
pyPDF

#env variable
python-dotenv

#tyoing and utilities
pydantic
requests 
tqdm

#logging
loguru

# uv pip install -r requirements.txt